refactor: 거래 도메인 구조 개선
