+++ src/main/java/com/ureca/snac/trade/service/AttachmentServiceImpl.java
@@ -6,14 +6,13 @@ import com.ureca.snac.common.s3.S3Path;
 import com.ureca.snac.common.s3.S3Uploader;
 import com.ureca.snac.member.Member;
 import com.ureca.snac.member.MemberRepository;
-import com.ureca.snac.trade.dto.AttachmentRequestDto;
-import com.ureca.snac.trade.dto.AttachmentResponseDto;
 import com.ureca.snac.trade.entity.Attachment;
 import com.ureca.snac.trade.entity.Trade;
 import com.ureca.snac.trade.exception.AttachmentAlreadyExistsException;
 import com.ureca.snac.trade.exception.AttachmentPermissionDenied;
 import com.ureca.snac.trade.repository.AttachmentRepository;
 import com.ureca.snac.trade.repository.TradeRepository;
+import com.ureca.snac.trade.service.interfaces.AttachmentService;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
