--- src/main/java/com/ureca/snac/trade/service/BasicTradeServiceImpl.java
@@ -1,292 +1,0 @@
-package com.ureca.snac.trade.service;
-
-import com.ureca.snac.auth.service.SnsService;
-import com.ureca.snac.board.entity.Card;
-import com.ureca.snac.board.entity.constants.CardCategory;
-import com.ureca.snac.board.entity.constants.SellStatus;
-import com.ureca.snac.board.exception.CardAlreadyTradingException;
-import com.ureca.snac.board.exception.CardInvalidStatusException;
-import com.ureca.snac.board.exception.CardNotFoundException;
-import com.ureca.snac.board.repository.CardRepository;
-import com.ureca.snac.member.Member;
-import com.ureca.snac.member.MemberRepository;
-import com.ureca.snac.member.exception.MemberNotFoundException;
-import com.ureca.snac.trade.controller.request.ClaimBuyRequest;
-import com.ureca.snac.trade.controller.request.CreateTradeRequest;
-import com.ureca.snac.trade.dto.TradeSide;
-import com.ureca.snac.trade.entity.Trade;
-import com.ureca.snac.trade.entity.TradeStatus;
-import com.ureca.snac.trade.exception.*;
-import com.ureca.snac.trade.repository.TradeRepository;
-import com.ureca.snac.trade.service.response.ProgressTradeCountResponse;
-import com.ureca.snac.trade.service.response.ScrollTradeResponse;
-import com.ureca.snac.trade.service.response.TradeResponse;
-import com.ureca.snac.wallet.Repository.WalletRepository;
-import com.ureca.snac.wallet.entity.Wallet;
-import com.ureca.snac.wallet.exception.WalletNotFoundException;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.util.List;
-
-import static com.ureca.snac.board.entity.constants.SellStatus.*;
-import static com.ureca.snac.trade.entity.TradeStatus.DATA_SENT;
-import static com.ureca.snac.trade.entity.TradeStatus.PAYMENT_CONFIRMED;
-
-@Slf4j
-@Service
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-public class BasicTradeServiceImpl implements BasicTradeService {
-
-    private final TradeRepository tradeRepository;
-    private final MemberRepository memberRepository;
-    private final CardRepository cardRepository;
-    private final WalletRepository walletRepository;
-
-    private final AttachmentService attachmentService;
-    private final SnsService snsService;
-
-    @Override
-    @Transactional
-    public Long createSellTrade(CreateTradeRequest createTradeRequest, String username) {
-        return buildTrade(createTradeRequest, username, SELLING);
-    }
-
-    @Override
-    @Transactional
-    public Long createBuyTrade(CreateTradeRequest createTradeRequest, String username) {
-        return buildTrade(createTradeRequest, username, PENDING);
-    }
-
-    @Override
-    @Transactional
-    public void cancelTrade(Long tradeId, String username) {
-        Trade trade = findLockedTrade(tradeId);
-        Member member = findMember(username);
-        Wallet wallet = findLockedWallet(trade.getBuyer().getId());
-        Card card = findLockedCard(trade.getCardId());
-
-        trade.cancel(member);
-
-        int refundMoney = trade.getPriceGb() - trade.getPoint();
-        if (refundMoney > 0) wallet.depositMoney(refundMoney);
-        if (trade.getPoint() > 0) wallet.depositPoint(trade.getPoint());
-
-        cardRepository.delete(card);
-    }
-
-    @Override
-    @Transactional
-    public void acceptBuyRequest(ClaimBuyRequest claimBuyRequest, String username) {
-        Card card = findLockedCard(claimBuyRequest.getCardId());
-        Member seller = findMember(username);
-        Trade trade = tradeRepository
-                .findLockedByCardId(claimBuyRequest.getCardId())
-                .orElseThrow(TradeNotFoundException::new);
-
-        if (card.getSellStatus() != SELLING) {
-            throw new CardAlreadyTradingException();
-        }
-
-        if (trade.getBuyer() == seller) {
-            throw new TradeSelfRequestException();
-        }
-
-        trade.changeSeller(seller);
-        card.changeSellStatus(TRADING);
-
-        String message = buildTradeMessage(trade, card);
-//        snsService.sendSms(trade.getSeller().getPhone(), message);
-    }
-
-    @Override
-    @Transactional
-    public void sendTradeData(Long tradeId, String username, MultipartFile picture) {
-        Trade trade = findLockedTrade(tradeId);
-        Member seller = findMember(username);
-        Card card = findLockedCard(trade.getCardId());
-
-        if (trade.getStatus() != PAYMENT_CONFIRMED) {
-            throw new TradeInvalidStatusException();
-        }
-
-        if (trade.getBuyer() == seller)
-            throw new TradeSendPermissionDeniedException();
-
-        if (trade.getSeller() != seller) { // 이미 지정된 판매자가 현재 요청자가 아닐 경우 권한 없음
-            throw new TradeSendPermissionDeniedException();
-        }
-
-        // 파일 업로드
-        attachmentService.upload(tradeId, username, picture);
-
-        trade.changeStatus(DATA_SENT);
-
-        String message = buildTradeMessage(trade, card);
-//        snsService.sendSms(trade.getBuyer().getPhone(), message);
-    }
-
-    @Override
-    @Transactional
-    public void confirmTrade(Long tradeId, String username) {
-        Trade trade = findLockedTrade(tradeId);
-        Member buyer = findMember(username);
-        Wallet wallet = findLockedWallet(trade.getSeller().getId());
-        Card card = findLockedCard(trade.getCardId());
-
-        trade.confirm(buyer);
-        card.changeSellStatus(SOLD_OUT);
-        wallet.depositMoney(trade.getPriceGb() - trade.getPoint());
-
-        String message = buildTradeMessage(trade, card);
-//        snsService.sendSms(trade.getSeller().getPhone(), message);
-    }
-
-    @Override
-    public ScrollTradeResponse scrollTrades(String username, TradeSide side, int size, Long lastTradeId) {
-        Member member = findMember(username);
-
-        List<Trade> trades = (side == TradeSide.BUY)
-                ? tradeRepository.findTradesByBuyerInfinite(member.getId(), lastTradeId, size + 1)
-                : tradeRepository.findTradesBySellerInfinite(member.getId(), lastTradeId, size + 1);
-
-        boolean hasNext = trades.size() > size;
-
-        List<TradeResponse> dto = trades.stream()
-                .limit(size)
-                .map(t -> TradeResponse.from(t, side))
-                .toList();
-
-        return new ScrollTradeResponse(dto, hasNext);
-    }
-
-    @Override
-    public ProgressTradeCountResponse countSellingProgress(String username) {
-        Member seller = findMember(username);
-
-        return new ProgressTradeCountResponse(tradeRepository.countBySellerAndStatusIn(seller,
-                List.of(DATA_SENT, PAYMENT_CONFIRMED)));
-    }
-
-    @Override
-    public ProgressTradeCountResponse countBuyingProgress(String username) {
-        Member buyer = findMember(username);
-
-        return new ProgressTradeCountResponse(tradeRepository.countByBuyerAndStatusIn(buyer,
-                List.of(DATA_SENT, PAYMENT_CONFIRMED)));
-    }
-
-    // === private helper === //
-    private Long buildTrade(CreateTradeRequest createTradeRequest, String username, SellStatus requiredStatus) {
-        Member member = findMember(username);
-        Wallet wallet = findLockedWallet(member.getId());
-        Card card = findLockedCard(createTradeRequest.getCardId());
-
-        if (requiredStatus == SELLING && card.getSellStatus() != SELLING) {
-            throw new CardAlreadyTradingException();
-        }
-
-        ensureStatus(card, requiredStatus);
-        ensureOwnership(card, member, requiredStatus);
-
-        int totalPay = createTradeRequest.getMoney() + createTradeRequest.getPoint();
-
-        if (card.getPrice() != totalPay)
-            throw new TradePaymentMismatchException();
-
-        if (createTradeRequest.getMoney() != 0) {
-            wallet.withdrawMoney(createTradeRequest.getMoney());
-        }
-        if (createTradeRequest.getPoint() != 0) {
-            wallet.withdrawPoint(createTradeRequest.getPoint());
-        }
-
-        Trade trade = Trade.buildTrade(createTradeRequest.getPoint(), member, member.getPhone(), card, requiredStatus);
-        tradeRepository.save(trade);
-
-        // 카드 상태 변경 (판매글이면 TRADING, 구매글이면 SELLING)
-        card.changeSellStatus(requiredStatus == SELLING ? TRADING : SELLING);
-
-        if (requiredStatus == SELLING) {
-            String message = buildTradeMessage(trade, card);
-//            snsService.sendSms(trade.getSeller().getPhone(), message);
-        }
-
-        return trade.getId();
-    }
-
-    private Member findMember(String email) {
-        return memberRepository.findByEmail(email).orElseThrow(MemberNotFoundException::new);
-    }
-
-    private Wallet findLockedWallet(Long memberId) {
-        return walletRepository.findByMemberIdWithLock(memberId).orElseThrow(WalletNotFoundException::new);
-    }
-
-    private Card findLockedCard(Long cardId) {
-        return cardRepository.findLockedById(cardId).orElseThrow(CardNotFoundException::new);
-    }
-
-    private Trade findLockedTrade(Long tradeId) {
-        return tradeRepository.findLockedById(tradeId).orElseThrow(TradeNotFoundException::new);
-    }
-
-    private void ensureStatus(Card card, SellStatus sellStatus) {
-        if (card.getSellStatus() != sellStatus) {
-            throw new CardInvalidStatusException();
-        }
-    }
-
-    private void ensureOwnership(Card card, Member member, SellStatus requiredStatus) {
-        boolean isOwner = card.getMember().equals(member);
-
-        // 판매 요청 시 자기 글 요청 방지
-        if (requiredStatus == SELLING && isOwner) {
-            throw new TradeSelfRequestException();
-        }
-        // 구매 요청 시 타인의 글만 허용
-        if (requiredStatus == PENDING && !isOwner) {
-            throw new TradePermissionDeniedException();
-        }
-    }
-
-
-    private String buildTradeMessage(Trade trade, Card card) {
-        TradeStatus status = trade.getStatus();
-
-        String statusText;
-        if (status == TradeStatus.PAYMENT_CONFIRMED && card.getCardCategory() == CardCategory.BUY && card.getSellStatus() == TRADING) {
-            statusText = "구매글 거래가 시작되었습니다";
-        } else {
-            switch (status) {
-                case PAYMENT_CONFIRMED:
-                    statusText = "결제가 완료되었습니다";
-                    break;
-                case DATA_SENT:
-                    statusText = "데이터 전송이 완료되었습니다";
-                    break;
-                case COMPLETED:
-                    statusText = "거래가 확정되었습니다";
-                    break;
-                case CANCELED:
-                    statusText = "거래가 취소되었습니다";
-                    break;
-                default:
-                    statusText = "상태가 변경되었습니다";
-            }
-        }
-
-        // 데이터량 GB, 가격 원 단위 포맷
-        int dataGb = trade.getDataAmount();
-        String priceText = String.format("%,d원", trade.getPriceGb());
-
-        return String.format(
-                "[SNAC] 거래 ID %d: %s.\n 데이터 %dGB, 거래 금액 %s.\n",
-                trade.getId(), statusText, dataGb, priceText
-        );
-    }
-}
