+++ src/main/java/com/ureca/snac/trade/service/interfaces/TradeStatisticsService.java
@@ -1,4 +1,4 @@
-package com.ureca.snac.trade.service;
+package com.ureca.snac.trade.service.interfaces;
 
 import com.ureca.snac.board.entity.constants.Carrier;
 import com.ureca.snac.trade.service.response.TradeStatisticsResponse;
